Fix warning: -Wnarrowing
Reported by GCC:
warning: narrowing conversion of ‘HeeksColor(187, 233, 255).HeeksColor::COLORREF_color()’ from ‘long int’ to ‘int’ inside { } is ill-formed in C++11 [-Wnarrowing]

According to C++ standards, long int ensures to have a least 32 bits wide integers, while int only ensures 16 bits.
As we need 24 bits to stores RBG (3x8 bits), we should ensures RGB variables to have 32 bits (ie. long int)

+++ trunk/src/HeeksCAD.cpp
@@ -330,7 +330,7 @@ bool HeeksCADapp::OnInit()
 	config.Read(_T("DrawGrid"), &digitizing_grid);
 	config.Read(_T("DrawRadius"), &digitizing_radius);
 	{
-		int default_color[NUM_BACKGROUND_COLORS] = {
+		long default_color[NUM_BACKGROUND_COLORS] = {
 			HeeksColor(187, 233, 255).COLORREF_color(),
 			HeeksColor(255, 255, 255).COLORREF_color(),
 			HeeksColor(247, 198, 243).COLORREF_color(),
